# api_go/cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/account/organization": {
            "post": {
                "summary": "Create an Organization",
                "parameters": [
                    {
                        "description": "Create organization",
                        "name": "organization",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.organizationCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.OrganizationModel"
                        }
                    }
                }
            }
        },
        "/v1/account/organization/{id}": {
            "get": {
                "summary": "Get an Organization",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Organization Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.OrganizationModel"
                        }
                    }
                }
            }
        },
        "/v1/account/user": {
            "post": {
                "summary": "Create a user",
                "parameters": [
                    {
                        "description": "Create user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.UserModel"
                        }
                    }
                }
            }
        },
        "/v1/account/user/{id}": {
            "get": {
                "summary": "Get an user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.UserModel"
                        }
                    }
                }
            }
        },
        "/v1/health": {
            "get": {
                "description": "Get component by component health status",
                "summary": "Get Health",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.healthResponse"
                        }
                    }
                }
            }
        },
        "/v1/participant/{id}": {
            "get": {
                "summary": "Get an participant",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.ParticipantModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Agreement": {
            "type": "object",
            "properties": {
                "acceptedTime": {
                    "type": "string"
                },
                "contractVersionId": {
                    "type": "string"
                },
                "participantId": {
                    "type": "string"
                },
                "permissionId": {
                    "type": "string"
                }
            }
        },
        "domain.OrganizationMember": {
            "type": "object",
            "properties": {
                "role": {
                    "$ref": "#/definitions/domain.Role"
                }
            }
        },
        "domain.OrganizationModel": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "members": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/domain.OrganizationMember"
                    }
                },
                "name": {
                    "type": "string"
                },
                "removed": {
                    "type": "boolean"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
        "domain.ParticipantModel": {
            "type": "object",
            "properties": {
                "acceptedPermissionIds": {
                    "description": "read model. Calculated from AllAgreements",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "allAgreements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Agreement"
                    }
                },
                "created": {
                    "type": "string"
                },
                "externalId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "organizationId": {
                    "type": "string"
                },
                "removed": {
                    "type": "boolean"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
        "domain.Role": {
            "type": "object"
        },
        "domain.User": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "domain.UserModel": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "removed": {
                    "type": "boolean"
                },
                "updated": {
                    "type": "string"
                }
            }
        },
        "server.componentHealth": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "isHealthy": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "server.healthResponse": {
            "type": "object",
            "properties": {
                "components": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.componentHealth"
                    }
                }
            }
        },
        "server.organizationCreateRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        }
    }
}